feat: Add utility functions for formatting time and retrieving prayer times
