+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/main/MainViewModel.kt
@@ -9,10 +9,10 @@ import net.thechance.mena.faith.domain.entity.PrayerTime
 import net.thechance.mena.faith.domain.repository.PrayerTimeRepository
 import net.thechance.mena.faith.domain.repository.QuranRepository
 import net.thechance.mena.faith.presentation.base.BaseViewModel
-import net.thechance.mena.faith.presentation.feature.main.MainMapper.getHijriDate
-import net.thechance.mena.faith.presentation.feature.main.MainMapper.getSunriseTime
 import net.thechance.mena.faith.presentation.feature.main.MainMapper.toTilawahUiState
 import net.thechance.mena.faith.presentation.feature.main.MainMapper.toUi
+import net.thechance.mena.faith.presentation.util.extentions.getHijriDate
+import net.thechance.mena.faith.presentation.util.extentions.getSunriseTime
 import kotlin.time.Clock
 import kotlin.time.ExperimentalTime
