+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/feature/main/componant/TilawahSection.kt
@@ -40,8 +40,7 @@ fun TilawahSection(
     modifier: Modifier
 ) {
     Column(
-        modifier = modifier
-            .fillMaxWidth(),
+        modifier = modifier.fillMaxWidth(),
         verticalArrangement = Arrangement.spacedBy(Theme.spacing._12)
     ) {
         Text(
