+++ faith_presentation/src/commonMain/kotlin/net/thechance/mena/faith/presentation/util/extentions/GetHijriDate.kt
@@ -1,0 +1,6 @@
+package net.thechance.mena.faith.presentation.util.extentions
+
+import net.thechance.mena.faith.domain.entity.PrayerTime
+
+fun getHijriDate(prayerTimes: List<PrayerTime>): String =
+    prayerTimes.firstOrNull()?.hijriDate ?: ""
